chore: consolidate Flask CLI commands into dedicated cli.py and streamline app setup
- Move all custom CLI commands (init‑db, seed‑labs, create‑admin, convert‑quantities, update‑lab‑codes) into `app/cli.py`
- Register commands via `init_cli(app)` in `create_app()` instead of importing `manage.py`
- Simplify `manage.py` to only run the Flask app
- Remove circular imports and ensure commands appear in `flask --help`

+++ app/models/transfer_log.py
@@ -5,7 +5,7 @@ from app.extensions import db
 
 class TransferLog(db.Model):
     id = db.Column(db.Integer, primary_key=True)
-    quantity = db.Column(db.Float, nullable=False)
+    quantity = db.Column(db.Integer, nullable=False)
     timestamp = db.Column(db.DateTime, default=datetime.utcnow, nullable=False)
     notes = db.Column(db.Text)
